backend/requirements.txt
fastapi==0.135.1
uvicorn==0.42.0
asyncpg==0.30.0
google-genai==1.69.0
python-dotenv==1.2.2
pydantic==2.12.5
httpx==0.28.1
slowapi==0.1.9
